New library: pmdmini
pmdmini is a library for playback of the PMD format used by the Japanese
PC-9801 computer.
This should be a complete implementation of the format, with correct
playback time, title/composer support, etc. It was tested on the iPhone
Simulator.

Possible TODO:
pmdmini has a nifty note retrieval function. See whether this can be
used to make the pianoroll work.

Integrates code from BouKiCHi's GPL pmdmini library, from:
https://github.com/BouKiCHi/mdxplayer
pmdmini is based on Koji AGAWA's pmdxmms, which in turn was based on the
standard PMDWin library by C60.