Initial support for 22khz DSF audio
DSF output now defaults to 22050hz. This is not selectable in the UI,
but there is a flag for it that should be hooked up at some point.
These changes allow most DSF files to play back at full speed on the
iPhone 4. (Some are still choppy, at least 80% of the hundreds of
tracks in Shenmue 1 and 2 play back full speed)